Added a new file

+++ 0x00-hello_world/6-size.c
@@ -22,7 +22,7 @@ int main(void)
 	/*
 	 * Display the size of a long long int in bytes
 	 */
-	printf("Size of a long long int: %d byte(s)\n", sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(long long int));
 	/*
 	 * Display the size of a float in byte
 	 */
